requirements/base.txt
# Bleeding edge Django
django

# Configuration
django-configurations
django-secure
django-cache-url
dj-database-url

# Forms
django-braces

# Models
django-model-utils

# For the persistance stores
psycopg2

# Unicode slugification
unicode-slugify
django-autoslug

# Time zones support
pytz

# Jinja2
Jinja2
jingo
jingo-markdown

# django-storages is a collection of custom storage backends for Django
# django-storages

# Your custom requirements go here
num2words

django_countries
django-datetime-widget
